Rename: 폴더 구조 변경
랜딩 페이지 경로 더 깔끔하게 수정
