+++ pages/faq/index.html
@@ -9,7 +9,7 @@
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/static/pretendard-dynamic-subset.css" />
     <link rel="stylesheet" href="/styles/reset.css" />
     <link rel="stylesheet" href="/styles/common.css" />
-    <link rel="stylesheet" href="/styles/faq.css" />
+    <link rel="stylesheet" href="faq.css" />
     <title>Linkbrary</title>
   </head>
   <body>
