+++ pages/member/signin/index.html
@@ -38,9 +38,9 @@
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/static/pretendard-dynamic-subset.css" />
     <link rel="stylesheet" href="/styles/reset.css" />
     <link rel="stylesheet" href="/styles/common.css" />
-    <link rel="stylesheet" href="/styles/sign.css" />
+    <link rel="stylesheet" href="../sign.css" />
 
-    <script defer type="module" src="/scripts/signin.js"></script>
+    <script defer type="module" src="signin.js"></script>
   </head>
   <body class="bgc-blue">
     <main>
@@ -49,12 +49,12 @@
           <a href="/" aria-label="메인 페이지로 이동">
             <img
               class="logo__img"
-              src="./images/icons/logo.svg"
+              src="/images/icons/logo.svg"
               alt="링크브러리 로고" />
           </a>
           <p class="logo__signup">
             <span>회원이 아니신가요?</span>
-            <a class="logo__signup--link" href="/signup.html">회원 가입하기</a>
+            <a class="logo__signup--link" href="../signup/">회원 가입하기</a>
           </p>
         </div>
