+++ pages/member/signup/index.html
@@ -39,9 +39,9 @@
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/static/pretendard-dynamic-subset.css" />
     <link rel="stylesheet" href="/styles/reset.css" />
     <link rel="stylesheet" href="/styles/common.css" />
-    <link rel="stylesheet" href="/styles/sign.css" />
+    <link rel="stylesheet" href="../sign.css" />
 
-    <script defer type="module" src="/scripts/signup.js"></script>
+    <script defer type="module" src="signup.js"></script>
   </head>
   <body class="bgc-blue">
     <main>
@@ -50,12 +50,12 @@
           <a href="/" aria-label="메인 페이지로 이동">
             <img
               class="logo__img"
-              src="./images/icons/logo.svg"
+              src="/images/icons/logo.svg"
               alt="링크브러리 로고" />
           </a>
           <p class="logo__signup">
             <span>이미 회원이신가요?</span>
-            <a class="logo__signup--link" href="/signin.html">로그인 하기</a>
+            <a class="logo__signup--link" href="../signin/">로그인 하기</a>
           </p>
         </div>
